+++ packages/interactivity/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   Allow more iterables to be used in each directives ([#67798](https://github.com/WordPress/gutenberg/pull/67798)).
+
+### Bug Fixes
+
+-   Fix an error when the value used in an each directive is not iterable ([#67798](https://github.com/WordPress/gutenberg/pull/67798)).
+
 ## 6.14.0 (2024-12-11)
 
 ## 6.13.0 (2024-11-27)